[array api] return NamedTuple from np.linalg APIs
